+++ slangpy/tests/slangpy_tests/test_torchintegration.py
@@ -13,6 +13,8 @@ except ImportError:
 if sys.platform == "darwin":
     pytest.skip("PyTorch requires CUDA, that is not available on macOS", allow_module_level=True)
 
+pytest.skip("Skipped due to sync issues / race condition with CUDA", allow_module_level=True)
+
 TEST_CODE = """
 import tensor;
 [Differentiable]