+++ docs/changelog.rst
@@ -7,9 +7,11 @@ Changelog
 
 SlangPy uses a `semantic versioning <http://semver.org>`__ policy for its API.
 
-Version 0.27.0 (TBA)
+Version 0.27.0 (May 9, 2025)
 ----------------------------
 
+- Package and distribute pytest tests. Fix deploying ``.pyi`` files in wheels + other minor fixes.
+  (PR `#196 <https://github.com/shader-slang/slangpy/pull/197>`__).
 - Introduce basic support for bindless textures and samplers. Currently only supported on D3D12.
   Add ``sgl::Feature::bindless`` (``slangpy.Feature.bindless``) to detect bindless support.
   Add ``sgl::DescriptorHandle`` (``slangpy.DescriptorHandle``) to represent bindless descriptor handles.